+++ config/prettier.config.js
@@ -2,5 +2,5 @@ module.exports = {
     singleQuote: true,
     tabWidth: 4,
     arrowParens: 'avoid',
-    endOfLine: 'auto'
+    endOfLine: 'auto',
 };