+++ index.d.ts
@@ -16,22 +16,22 @@ declare module 'pc-nrfconnect-shared' {
             message: string;
         };
         device: {
-            devices: Device[];
+            devices: readonly Device[];
             deviceInfo: any;
             isSetupDialogVisible: boolean;
             isSetupWaitingForUserInput: boolean;
             selectedSerialNumber: string;
-            setupDialogChoices: string[];
+            setupDialogChoices: readonly string[];
             setupDialogText: string | null;
         };
         errorDialog: {
             errorResolutions: any;
             isVisible: boolean;
-            messages: string[];
+            messages: readonly string[];
         };
         log: {
             autoScroll: boolean;
-            logEntries: string[];
+            logEntries: readonly string[];
         };
     }
 
@@ -45,7 +45,7 @@ declare module 'pc-nrfconnect-shared' {
 
     interface LogAddEntries {
         type: NrfConnectActionType.LOG_ADD_ENTRIES;
-        entries: any[];
+        entries: readonly any[];
     }
 
     interface LogClearEntries {
@@ -105,7 +105,7 @@ declare module 'pc-nrfconnect-shared' {
          *
          * `[['Connection Map', ConnectionMap], ['Server Setup', ServerSetup]]`
          */
-        panes: [string, React.FC][];
+        panes: readonly (readonly [string, React.FC])[];
         /**
          * Describes whether the log will show automatically when the
          * application starts. Defaults to `true`.
@@ -162,7 +162,7 @@ declare module 'pc-nrfconnect-shared' {
     interface Device {
         boardVersion: string;
         serialNumber: string;
-        traits: string[];
+        traits: readonly string[];
         serialport: {
             path: string;
             manufacturer: string;
@@ -197,30 +197,31 @@ declare module 'pc-nrfconnect-shared' {
         deviceSetup?: DeviceSetup;
         /**
          * This callback is invoked before a device is about to be
-         * rogrammed. If no `deviceSetup `is provided, this callback
-         * will not be invoked.
+         * programmed. If no `deviceSetup `is provided, this callback
+         * will not be invoked. If this returns a promise, programming
+         * is only done after the promise is fulfilled.
          */
-        releaseCurrentDevice?: () => any;
+        releaseCurrentDevice?: () => Promise<void> | void;
         /**
          * This callback is invoked when a device is selected by
          * the user. The callback receives the selected device as
          * a parameter.
          */
-        onDeviceSelected?: (device: Device) => any;
+        onDeviceSelected?: (device: Device) => void;
         /**
          * This callback is invoked when programming a device is
          * finished. The callback receives the programmed device
          * as a parameter. If no `deviceSetup` is provided, this
          * callback will not be invoked.
          */
-        onDeviceIsReady?: (device: Device) => any;
+        onDeviceIsReady?: (device: Device) => void;
         /**
          * This callback is invoked when a selected device is again
          * deselected. This may be caused by the user deselecting
          * the device but also automatically if programming a device
          * failed.
          */
-        onDeviceDeselected?: () => any;
+        onDeviceDeselected?: () => void;
     }
 
     /**
@@ -255,8 +256,8 @@ declare module 'pc-nrfconnect-shared' {
         isVisible: boolean;
         title?: string;
         text?: string;
-        onOk: () => any;
-        onCancel: () => any;
+        onOk: () => void;
+        onCancel: () => void;
         okButtonText?: string;
         cancelButtonText?: string;
         isInProgress?: boolean;
@@ -275,13 +276,13 @@ declare module 'pc-nrfconnect-shared' {
 
     export class SliderProps {
         id?: string;
-        values: number[];
+        values: readonly number[];
         range: {
             min: number;
             max: number;
         };
-        onChange: ((value: number) => any)[];
-        onChangeComplete?: () => any;
+        onChange: readonly ((value: number) => void)[];
+        onChangeComplete?: () => void;
     }
 
     export class Slider extends React.Component<SliderProps> {}
@@ -289,7 +290,7 @@ declare module 'pc-nrfconnect-shared' {
     // ErrorDialog.jsx
 
     export interface ErrorDialogState {
-        messages: string[];
+        messages: readonly string[];
         isVisible: boolean;
         errorResolutions: any;
     }
@@ -318,7 +319,7 @@ declare module 'pc-nrfconnect-shared' {
     interface InlineInputProps {
         value: string;
         isValid?: (value: string) => boolean;
-        onChange: (value: string) => any;
+        onChange: (value: string) => void;
         className?: string;
         style?: React.CSSProperties;
     }
@@ -333,10 +334,12 @@ declare module 'pc-nrfconnect-shared' {
             min: number;
             max: number;
         };
-        onChange: (value: number) => any;
+        onChange: (value: number) => void;
     }
 
-    export class NumberInlineInput extends React.Component<NumberInlineInputProps> {}
+    export class NumberInlineInput extends React.Component<
+        NumberInlineInputProps
+    > {}
 
     // Toggle.jsx
 
@@ -386,10 +389,10 @@ declare module 'pc-nrfconnect-shared' {
     // systemReport.js
 
     export function systemReport(
-        allDevices: any,
-        currentSerialNumber: string[],
+        allDevices: readonly Device[],
+        currentSerialNumber: string,
         currentDevice: any
-    ): any;
+    ): Promise<void>;
 
     // userData.js