+++ Changelog.md
@@ -1,3 +1,13 @@
+## 4.9.5
+### Fixed
+- More type enhancements:
+  - Correct `systemReport` signature
+  - Limit array types to readonly where possible
+  - Do not use void as a callback return value
+- Make linting fail if the type check fails
+- Remove conflicting ESLint rule 'quotes'
+- Lint all JavaScript files in this project, not just the ones in src/
+
 ## 4.9.4
 ### Fixes
 - Fix app's dependencies in webpack config